proje_dosyalari/ai_adapter_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
# AI Model SDKs
openai
anthropic
mistralai
llama-cpp-python
huggingface_hub
transformers
# Utilities
python-dotenv
loguru
requests
# Testing
pytest
pytest-asyncio

